Chapter 3 Exercises 22.12

+++ Chapter_3/input_output_float.cpp
@@ -3,8 +3,8 @@
 using namespace std;
 int main()
 {
-	float a;
+	double a;
 	cin>>a;
-	printf("%.3lf", a);
+	printf("%.3lld", a);
 	return 0;
 }
